requirements.txt
blinker #==1.7.0
click #==8.1.7
Flask #==3.0.1
itsdangerous #==2.1.2
Jinja2 #==3.1.3
MarkupSafe #==2.1.4
num2word #==1.0.1
pyserial #==3.5
RPi.GPIO #==0.7.1
Werkzeug #==3.0.1
